dom/media/test/test_streams_autoplay.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test that a MediaStream source triggers autoplay</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
  <script type="text/javascript" src="manifest.js"></script>
</head>
<body>
<video id="v1"></video>
<video id="v2" autoplay></video>
<pre id="test">
<script class="testbody" type="text/javascript">
SimpleTest.waitForExplicitFinish();

var media = getPlayableVideo(gSmallTests);

if (media == null) {
  todo(false, "No media supported.");
  SimpleTest.finish();
} else {
  v1.src = media.name;
  v1.preload = 'metadata';
  v1.onloadedmetadata = function() {
    v2.srcObject = v1.mozCaptureStream();
    v1.play();
  };
  var onPlayingEventFired = false;

  v2.onplaying = function() {
    if(!onPlayingEventFired) {
      onPlayingEventFired = true;
      ok(true, "playback started");
      SimpleTest.finish();
    }
  }
}
</script>
</pre>
</body>
</html>
